requirements.txt
gmr==1.6.2
numpy==1.21.5
scipy==1.7.3
matplotlib==3.5.2
pyLasaDataset==0.1.1
casadi==3.6.3
cvxpy==1.3.2
dtw-python==1.3.0
mosek == 10.0.34
matlabengine